+++ resources/views/admin/slide/list.blade.php
@@ -40,7 +40,7 @@
                                         <img src="upload/slide/{{$sl->Hinh}}" width="400px" height="300px">
                                     </td>
                                     <td>{{$sl->link}}</td>
-                                    <td class="center"><i class="fa fa-trash-o  fa-fw"></i><a href="admin/slide/delete/{{$sl->id}}"> Delete</a></td>
+                                    <td class="center"><i class="fa fa-trash-o  fa-fw"></i><a href="admin/slide/delete/{{$sl->id}}" onclick="return confirm('Bạn có chắc muốn xoá?')"> Delete</a></td>
                                     <td class="center"><i class="fa fa-pencil fa-fw"></i> <a href="admin/slide/edit/{{$sl->id}}">Edit</a></td>
                                 </tr>
                             @endforeach
