+++ resources/views/admin/theloai/list.blade.php
@@ -34,7 +34,7 @@
                                     <td>{{$tl->id}}</td>
                                     <td>{{$tl->Ten}}</td>
                                     <td>{{$tl->TenKhongDau}}</td>
-                                    <td class="center"><i class="fa fa-trash-o  fa-fw"></i><a href="admin/theloai/delete/{{$tl->id}}"> Delete</a></td>
+                                    <td class="center"><i class="fa fa-trash-o  fa-fw"></i><a href="admin/theloai/delete/{{$tl->id}}" onclick="return confirm('Bạn có chắc muốn xoá?')"> Delete</a></td>
                                     <td class="center"><i class="fa fa-pencil fa-fw"></i> <a href="admin/theloai/edit/{{$tl->id}}">Edit</a></td>
                                 </tr>
                             @endforeach
